requirements.txt
torch
torch-geometric
networkx
matplotlib
umap-learn
scikit-learn
numpy
pandas